Move contents of many plugin headers to corresponding source files in TopKinFitter
List of classes:
- TtFullHadKinFitProducer
- TtFullLepKinSolutionProducer
- TtSemiLepKinFitProducer
